requirements.txt
FastAPI
LangGraph
uvicorn
jinja2
httpx
python-dotenv


langchain
langchain-openai
python-dotenv